+++ src/main/kotlin/io/snyk/plugin/settings/SnykProjectSettingsConfigurable.kt
@@ -45,7 +45,7 @@ class SnykProjectSettingsConfigurable(val project: Project) : SearchableConfigur
         }
 
         applicationSettingsStateService.customEndpointUrl = customEndpoint
-        SnykCodeParams.instance.apiUrl = toSnykCodeApiUrl(customEndpoint)
+        SnykCodeParams.instance.apiUrl = customEndpoint
         SnykCodeParams.instance.isDisableSslVerification = snykSettingsDialog.isIgnoreUnknownCA()
 
         applicationSettingsStateService.token = snykSettingsDialog.getToken()